Extract rule to separate file
