+++ cell-autom/CMakeLists.txt
@@ -3,6 +3,7 @@ include_directories(${PROJECT_SOURCE_DIR}/cell-autom)
 
 add_library(
     cell-autom
-    field.cpp
     cell.cpp
+    field.cpp
+    rule.cpp
 )
