+++ cell-autom/field.cpp
@@ -1,7 +1,8 @@
 #include <field.h>
 
 #include <cassert>
-#include <algorithm>
+
+#include <rule.h>
 
 namespace {
 
@@ -76,32 +77,6 @@ Cell Field::cell(CellIndex index) const {
 
 Cell Field::cellNextGeneration(CellIndex index) const {
     return rule(cell(index), neighbours(index));
-}
-
-size_t aliveCount(Cells neighbours) {
-    return std::count(
-            std::begin(neighbours),
-            std::end(neighbours),
-            Cell::Alive
-    );
-
-}
-
-Cell rule(Cell cell, Cells neighbours) {
-    auto aliveAround = aliveCount(neighbours);
-
-    switch (cell) {
-        case Cell::Died:
-            return aliveAround == 2
-                ? Cell::Alive
-                : Cell::Died;
-        case Cell::Alive: 
-            return (aliveAround == 2) || (aliveAround == 3)
-                ? Cell::Alive
-                : Cell::Died;
-        default:
-            assert(false && "Unknown Cell::<enumerator>");
-    };
 }
 
 Field nextGeneration(const Field& current) {
