+++ cell-autom/rule.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include <cell.h>
+#include <cells.h>
+
+Cell rule(Cell cell, Cells neighbours);
